backend/cerebro_fetcher/requirements.txt
requests>=2.31.0
pandas>=2.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0
google-cloud-bigquery>=3.11.0
db-dtypes>=1.1.1
